requirements.txt
langchain
openai
duckduckgo-search
tiktoken
python-dotenv
langchain-community
langchain-deepseek
langchain-anthropic
tavily-python